requirements_api.txt
# Django REST Framework for API
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Advanced ID Card System Dependencies
Pillow>=10.0.0
qrcode[pil]>=7.4.2
pandas>=2.0.0
openpyxl>=3.1.0
reportlab>=4.0.0# Optional: OCR and AI Features (install separately if needed)
# easyocr>=1.7.0  # For OCR text extraction from ID cards
# opencv-python>=4.8.0  # For image processing
# deepface>=0.0.79  # For facial verification (requires TensorFlow)
# tensorflow>=2.13.0  # Required for DeepFace